refactor(config): extract manifest diagnostics
